Test-commit-revert scripts
Watch runs inotifywait to check for file changes. It runs
the tcr script, which uses all the other "implementation"
scripts. Changed the revert to only rollback the
production code -- wanted to see watch in tmux pane while I
worked on the other files (afaict it is a necessity), but
it is impossible to save both files at once(?) This is better
anyways, rewriting the test is bordering on tedium, though
I can see the argument that you need to change the test to
chew off a smaller bite. You have to be honest with vim,
hitting undo (u) in the production code brings it back up. Also
a drag to hit :e to refresh, getting tired of research into
this though.

This was a good exercise in scripting, it feels good to have
something automated, makes you want to learn more. The output
of the revisions in teh gist is awesome, micro-commits are
very readable.